server/requirements.txt
pyserial==3.5
websockets==10.3